requirements.txt
certifi==2020.6.20
dj-database-url==0.5.0
Django==2.2
djangorestframework==3.12.1
python-dotenv==0.15.0
pytz==2020.1
sqlparse==0.4.1
wincertstore==0.2
